Created tests for module1, module2 and removed the web project

+++ module1/src/main/java/com/geeth/modules/module1/Model1Repository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface Model1Repository extends JpaRepository<Model1, Long> {
 	
-	Model1 findById(int id);
+	Model1 findById(Long id);
 
 }
