+++ module2/src/main/java/com/geeth/modules/module2/Model2Repository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface Model2Repository extends JpaRepository<Model2, Long> {
 	
-	Model2 findById(int id);
+	Model2 findById(Long id);
 
 }
